# assistant-srv/api_docs.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Personal AI Assistant Server API",
...
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/oauth/{provider}/authorize": {
      "get": {
        "tags": [
          "oauth"
        ],
        "summary": "Get Oauth Authorize Url",
        "description": "Get OAuth authorization URL for the specified provider.",
        "operationId": "get_oauth_authorize_url_api_oauth__provider__authorize_get",
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Provider"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/oauth/{provider}/callback": {
      "post": {
        "tags": [
          "oauth"
        ],
        "summary": "Oauth Callback",
        "description": "Handle OAuth callback and login/register user.",
        "operationId": "oauth_callback_api_oauth__provider__callback_post",
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Provider"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OAuthCallbackRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthLoginResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/oauth/{provider}/unlink/{user_id}": {
      "post": {
        "tags": [
          "oauth"
        ],
        "summary": "Unlink Oauth Provider",
        "description": "Unlink OAuth provider from user account.",
        "operationId": "unlink_oauth_provider_api_oauth__provider__unlink__user_id__post",
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Provider"
            }
          },
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OAuthCallbackRequest": {
        "properties": {
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          }
        },
        "type": "object",
        "required": [
          "provider",
          "code"
        ],
        "title": "OAuthCallbackRequest",
        "description": "OAuth callback request."
      },
      "OAuthLoginResponse": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "is_new_user": {
            "type": "boolean",
            "title": "Is New User"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "username",
          "email",
          "display_name",
          "avatar_url",
          "role",
          "status",
          "is_new_user"
        ],
        "title": "OAuthLoginResponse",
        "description": "OAuth login response."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}